+++ lapce-data/src/plugin.rs
@@ -10,7 +10,7 @@ use druid::{
 use indexmap::IndexMap;
 use lapce_core::directory::Directory;
 use lapce_proxy::plugin::{download_volt, volt_icon, wasi::find_all_volts};
-use lapce_rpc::plugin::{VoltInfo, VoltMetadata};
+use lapce_rpc::plugin::{VoltID, VoltInfo, VoltMetadata};
 use parking_lot::Mutex;
 use plugin_install_status::PluginInstallStatus;
 use serde::{Deserialize, Serialize};
@@ -48,8 +48,8 @@ impl VoltIconKind {
 pub struct VoltsList {
     pub tab_id: WidgetId,
     pub total: usize,
-    pub volts: IndexMap<String, VoltInfo>,
-    pub icons: im::HashMap<String, VoltIconKind>,
+    pub volts: IndexMap<VoltID, VoltInfo>,
+    pub icons: im::HashMap<VoltID, VoltIconKind>,
     pub status: PluginLoadStatus,
     pub loading: Arc<Mutex<bool>>,
     pub event_sink: ExtEventSink,
@@ -143,12 +143,12 @@ pub struct PluginData {
     pub uninstalled_id: WidgetId,
 
     pub volts: VoltsList,
-    pub installing: IndexMap<String, PluginInstallStatus>,
-    pub installed: IndexMap<String, VoltMetadata>,
-    pub installed_latest: IndexMap<String, VoltInfo>,
-    pub installed_icons: im::HashMap<String, VoltIconKind>,
-    pub disabled: HashSet<String>,
-    pub workspace_disabled: HashSet<String>,
+    pub installing: IndexMap<VoltID, PluginInstallStatus>,
+    pub installed: IndexMap<VoltID, VoltMetadata>,
+    pub installed_latest: IndexMap<VoltID, VoltInfo>,
+    pub installed_icons: im::HashMap<VoltID, VoltIconKind>,
+    pub disabled: HashSet<VoltID>,
+    pub workspace_disabled: HashSet<VoltID>,
 }
 
 #[derive(Clone, PartialEq, Eq)]
@@ -167,8 +167,8 @@ pub struct PluginsInfo {
 impl PluginData {
     pub fn new(
         tab_id: WidgetId,
-        disabled: Vec<String>,
-        workspace_disabled: Vec<String>,
+        disabled: Vec<VoltID>,
+        workspace_disabled: Vec<VoltID>,
         event_sink: ExtEventSink,
     ) -> Self {
         {
@@ -210,11 +210,11 @@ impl PluginData {
         });
     }
 
-    pub fn plugin_disabled(&self, id: &str) -> bool {
+    pub fn plugin_disabled(&self, id: &VoltID) -> bool {
         self.disabled.contains(id) || self.workspace_disabled.contains(id)
     }
 
-    pub fn plugin_status(&self, id: &str) -> PluginStatus {
+    pub fn plugin_status(&self, id: &VoltID) -> PluginStatus {
         if self.plugin_disabled(id) {
             return PluginStatus::Disabled;
         }