+++ lapce-ui/src/tab.rs
@@ -995,7 +995,7 @@ impl LapceTab {
                     LapceUICommand::LoadPluginIcon(id, icon) => {
                         ctx.set_handled();
                         let plugin = Arc::make_mut(&mut data.plugin);
-                        plugin.volts.icons.insert(id.to_string(), icon.clone());
+                        plugin.volts.icons.insert(id.clone(), icon.clone());
                     }
                     LapceUICommand::VoltInstalled(volt, icon) => {
                         let plugin = Arc::make_mut(&mut data.plugin);
@@ -1190,10 +1190,9 @@ impl LapceTab {
                             .map(|diff| {
                                 let checked = source_control
                                     .file_diffs
-                                    .iter()
-                                    .find_map(|(p, c)| (p == &diff).then_some(*c))
-                                    .unwrap_or(true);
-                                (diff, checked)
+                                    .get(diff.path())
+                                    .map_or(true, |(_, c)| *c);
+                                (diff.path().clone(), (diff, checked))
                             })
                             .collect();