+++ CHANGELOG.md
@@ -7,14 +7,16 @@
 - [#1820](https://github.com/lapce/lapce/pull/1820): Add remote svg icon colour to theme, disable plugin settings when none are available
 
 ### Features/Changes
+- [#1899](https://github.com/lapce/lapce/pull/1899): Improve sorting files with numbers
 - [#1831](https://github.com/lapce/lapce/pull/1831): Plugin settings shown on right click
-
 - [#1830](https://github.com/lapce/lapce/pull/1830): Adds Clojure language support
-
 - [#1835](https://github.com/lapce/lapce/pull/1835): Add mouse keybinds
+- [#1856](https://github.com/lapce/lapce/pull/1856): Highlight git/VCS modified files in explorer, palette, and buffer tabs
+- [#1574](https://github.com/lapce/lapce/pull/1574): Panel sections can be expanded/collapsed
 - [#1938](https://github.com/lapce/lapce/pull/1938): Use dropdown for theme selection in settings
 
 ### Bug Fixes
+- [#1911](https://github.com/lapce/lapce/pull/1911): Fix movement on selections with left/right arrow keys
 
 ## 0.2.5